requirements.txt
langchain==0.3.21
langchain-core==0.3.46
langchain-experimental==0.3.4
langchain-community==0.3.20
langchain-openai==0.3.9
langchain-anthropic==0.3.10
langchain-text-splitters==0.3.7
langchain-elasticsearch==0.3.2
langchain-chroma==0.2.2
langchain-cohere==0.4.3
langchain-milvus==0.1.8
langchain-google-genai==2.1.0
langchain-huggingface==0.1.2
langchain-azure-ai==0.1.2
langchain-teddynote==0.3.44
langchainhub==0.1.21
langgraph==0.3.18
langsmith==0.3.18
huggingface-hub==0.29.3
openai==1.67.0
deepl==1.21.1
kiwipiepy==0.20.4
konlpy==0.6.0

pandas==2.2.3
rank-bm25==0.2.2

redis==5.2.1
chromadb==0.6.3

pymupdf==1.25.4
pypdf==4.3.1
pdfplumber==0.11.5
pdfminer-six==20231228
pymupdf4llm==0.0.17

matplotlib==3.10.1
streamlit==1.43.2
jupyter==1.1.1
notebook==7.3.3

torch==2.6.0
torchvision==0.21.0
faiss-cpu==1.10.0
open-clip-torch==2.31.0

python-dotenv==1.0.1
pydantic==2.10.6
lxml==5.3.1
pillow==10.4.0
lark==1.2.2
ragas==0.2.14
unstructured==0.17.0
arxiv==2.1.3
tiktoken==0.9.0
tenacity==8.5.0
pymilvus==2.5.5
google-search-results==2.4.2
protobuf==3.20.3
sqlalchemy==2.0.39
llama-index-core==0.11.23
llama-parse==0.6.4.post1
llama-index-readers-file==0.2.2
flashrank==0.2.10
docx2txt==0.8
nest-asyncio==1.6.0
#rapidocr-onnxruntime==1.4.4
seaborn==0.13.2
grandalf==0.8
rouge-score==0.1.2
langchain-ollama==0.2.3
mypy==1.15.0
pinecone==5.4.2
wikipedia==1.4.0
scikit-learn==1.6.1
gTTS==2.5.1